llm/requirements.txt
# Dependencies for LLM Policy Generator
# Pourquoi : Bibliothèques pour intégrer avec les LLMs et traiter les données
# Comment : pip install -r llm/requirements.txt

# DeepSeek R1 et Hugging Face API (utilise openai pour DeepSeek - API compatible)
openai>=1.3.0

# Hugging Face (pour chargement local optionnel - nécessite GPU)
# transformers>=4.35.0  # Optionnel - seulement si vous utilisez HuggingFace local
# torch>=2.1.0  # Optionnel - seulement si vous utilisez HuggingFace local
# accelerate>=0.24.0  # Optionnel

# Requêtes HTTP (pour Hugging Face API)
requests>=2.31.0

# Utilitaires
python-dotenv>=1.0.0  # Pour les variables d'environnement (API keys) - Optionnel (fallback manuel disponible)

# Traitement de texte (optionnel pour futures fonctionnalités)
# nltk>=3.8.1

